1) Run tests under firebase emulators
