+++ .gitignore
@@ -18,9 +18,15 @@
 .env.test.local
 .env.production.local
 
+# misc debug logs
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+# firebase debug logs
+database-debug.log
 firebase-debug.log*
+firestore-debug.log
+pubsub-debug.log
 
 .firebase
